refactor: move thumbnail information to datamodel
move thumbnail types to shared datamodel
remove ineffective scope rules (to be replaced)
